Correct merge resolution
The merge of sfc in "f6c30f3 Merge branch 'SLE11-SP2' into SLE11-SP3" favored
a version of the driver that looks like the one from -stable but, following
bnc#786035, the driver version in SP3 looks more like the one from v3.8-rc1.
The little fixups in this commit revert to the latter alternative.

- sfc: Really disable flow control while flushing (bnc#786035
  FATE#314299).
- Refresh
  patches.drivers/sfc-0084-sfc-Add-SR-IOV-back-end-support-for-SFC9000-family.patch.