kabi/severities: allow all kABI changes to pass, while still keeping SP3 symsets
We try to keep SLE11-SP4 kABI as stable as possible. This allows us
to keep track of how kABI is changing by watching the kABI build warnings
but do not have failing builds because of that.

This will be removed once kABI is officially frozen and we start tracking
it properly in our normal way.