Remove the temporary solution for save printk in NMI (bnc#831949)
The most critical situation has been solved upstream by the commit
a9edc88093287183ac9 ("x86/nmi: Perform a safe NMI stack trace on all CPUs").
It handles the scenarios reported by customers and should be enough.

The complex temporary solution, used on SLE12-GM and SLE12-SP1, is not worth
porting. It would be needed only for the dozen of remaining messages
printed from NMI context.

There is an alternative solution for the remaining printk's. It has a good
chance to get accepted but it has not happened yet. See,
https://lkml.kernel.org/g/<1459353210-20260-1-git-send-email-pmladek@suse.com>

- Delete
  patches.fixes/printk-nmi-0002-printk-allow-to-handle-more-log-buffers.patch.
- Delete
  patches.fixes/printk-nmi-0004-printk-add-NMI-ring-and-cont-buffers.patch.
- Delete
  patches.fixes/printk-nmi-0005-printk-allow-to-modify-NMI-log-buffer-size-using-boo.patch.
- Delete
  patches.fixes/printk-nmi-0006-printk-NMI-safe-printk.patch.
- Delete
  patches.fixes/printk-nmi-0007-printk-right-ordering-of-the-cont-buffers-from-NMI-c.patch.
- Delete
  patches.fixes/printk-nmi-0008-printk-try-hard-to-print-Oops-message-in-NMI-context.patch.
- Delete
  patches.fixes/printk-nmi-0009-printk-merge-and-flush-NMI-buffer-predictably-via-IR.patch.
- Delete
  patches.fixes/printk-nmi-0010-printk-survive-rotation-of-sequence-numbers.patch.
- Delete
  patches.fixes/printk-nmi-0011-printk-avoid-staling-when-merging-NMI-log-buffer.patch.
- Delete
  patches.fixes/printk-nmi-correct-detection-of-the-continuous-buffer-si.patch.