Use make --output-sync feature when available (bsc#1012422).
The mesages in make output can interleave making it impossible to
extract warnings reliably. Since version 4 GNU Make supports
--output-sync flag that prints output of each sub-command atomically
preventing this issue. Detect the flag and use it if available.

SLE11 has make 3.81 so it is required to include make 4 in the kernel
OBS projects to take advantege of this.