Delete patches.suse/kgr-0107-reserve-a-place-in-thread_struct-for-storing-RIP.patch.
A place to store a RIP was reserved in thread struct in order to be able
to patch __schedule() function with kGraft. This still has not been
fully implemented despite there is a prototype. Meanwhile the original
problem was solved upstream by rewriting switch_to() macro to a proper C
function. RIP is automatically stored there. It was merged to 4.9, so
we're gonna benefit from the change in SLE15. We made a decision not to
pursue this in SLE12 anymore.

Drop the patch.